+++ packages/components-js/tests/e2e/helpers/playwright-helper.ts
@@ -17,7 +17,7 @@ export type ClickableTests = {
 }[];
 
 // TODO: fix typing
-type Options = any & {
+export type Options = any & {
   enableLogging?: boolean;
   injectIntoHead?: string;
   withoutLoadCall?: boolean;