+++ packages/components-js/tests/e2e/specs/carousel.e2e.ts
@@ -29,6 +29,7 @@ type InitOptions = {
   rewind?: boolean;
   activeSlideIndex?: number;
   skipLinkTarget?: string;
+  dir?: 'ltr' | 'rtl';
 };
 
 const initCarousel = (page: Page, opts?: InitOptions) => {
@@ -40,6 +41,7 @@ const initCarousel = (page: Page, opts?: InitOptions) => {
     rewind = true,
     activeSlideIndex,
     skipLinkTarget,
+    dir = 'ltr',
   } = opts || {};
 
   const slides = Array.from(Array(amountOfSlides))
@@ -57,6 +59,7 @@ const initCarousel = (page: Page, opts?: InitOptions) => {
     rewind === false ? 'rewind="false"' : '',
     activeSlideIndex ? `active-slide-index="${activeSlideIndex}"` : '',
     skipLinkTarget ? `skip-link-target="${skipLinkTarget}"` : '',
+    dir ? `dir="${dir}"` : '',
   ].join(' ');
 
   const content = `${focusableElementBefore}<p-carousel heading="Heading" ${attrs}>
@@ -67,6 +70,7 @@ const initCarousel = (page: Page, opts?: InitOptions) => {
 };
 
 const getHost = (page: Page) => page.locator('p-carousel');
+const getSplide = (page: Page) => page.locator('#splide');
 const getSlottedSlides = async (page: Page) => page.locator('p-carousel').locator('[slot^="slide-"]').all();
 const getSplideTrack = (page: Page) => page.locator('p-carousel .splide__track');
 const getSlides = async (page: Page) => page.getByRole('group').all();
@@ -1048,5 +1052,69 @@ test.describe('lifecycle', () => {
     expect(status.componentDidUpdate['p-button-pure'], 'componentDidUpdate: p-button-pure').toBe(3);
     expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(3);
     expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(5);
+  });
+});
+
+test.describe('rtl mode', () => {
+  test('should render with rtl mode enabled', async ({ page }) => {
+    await initCarousel(page, { dir: 'rtl' });
+    const splide = getSplide(page);
+    await expect(splide).toHaveClass(/splide--rtl/);
+  });
+
+  test('should move slides on prev button clicks', async ({ page }) => {
+    await initCarousel(page, { dir: 'rtl' });
+    const buttonPrev = getButtonPrev(page);
+    const [slide1, slide2, slide3] = await getSlides(page);
+
+    await isElementCompletelyInViewport(slide1);
+    await isElementNotInViewport(slide2);
+    await isElementNotInViewport(slide3);
+
+    await buttonPrev.click();
+    await waitForSlideToBeActive(slide3);
+    await isElementNotInViewport(slide1);
+    await isElementNotInViewport(slide2);
+    await isElementCompletelyInViewport(slide3);
+
+    await buttonPrev.click();
+    await waitForSlideToBeActive(slide2);
+    await isElementNotInViewport(slide1);
+    await isElementCompletelyInViewport(slide2);
+    await isElementNotInViewport(slide3);
+
+    await buttonPrev.click();
+    await waitForSlideToBeActive(slide1);
+    await isElementCompletelyInViewport(slide1);
+    await isElementNotInViewport(slide2);
+    await isElementNotInViewport(slide3);
+  });
+
+  test('should move slides on next button clicks', async ({ page }) => {
+    await initCarousel(page, { dir: 'rtl' });
+    const buttonNext = getButtonNext(page);
+    const [slide1, slide2, slide3] = await getSlides(page);
+
+    await isElementCompletelyInViewport(slide1);
+    await isElementNotInViewport(slide2);
+    await isElementNotInViewport(slide3);
+
+    await buttonNext.click();
+    await waitForSlideToBeActive(slide2);
+    await isElementNotInViewport(slide1);
+    await isElementCompletelyInViewport(slide2);
+    await isElementNotInViewport(slide3);
+
+    await buttonNext.click();
+    await waitForSlideToBeActive(slide3);
+    await isElementNotInViewport(slide1);
+    await isElementNotInViewport(slide2);
+    await isElementCompletelyInViewport(slide3);
+
+    await buttonNext.click();
+    await waitForSlideToBeActive(slide1);
+    await isElementCompletelyInViewport(slide1);
+    await isElementNotInViewport(slide2);
+    await isElementNotInViewport(slide3);
   });
 });