+++ packages/components-js/tests/e2e/specs/flyout.e2e.ts
@@ -15,10 +15,11 @@ import {
   skipInBrowsers,
   sleep,
   waitForStencilLifecycle,
+  type Options,
 } from '../helpers';
 import { Components } from '@porsche-design-system/components';
 
-const CSS_TRANSITION_DURATION = 600;
+const CSS_TRANSITION_DURATION = 600; // Corresponds to motionDurationLong
 const flyoutMinWidth = 320;
 
 const getHost = (page: Page) => page.locator('p-flyout');
@@ -49,7 +50,8 @@ const initBasicFlyout = (
   other?: {
     markupBefore?: string;
     markupAfter?: string;
-  }
+  },
+  options?: Options
 ): Promise<void> => {
   const { header = '', content = '<p>Some Content</p>', footer = '', subFooter = '' } = flyoutSlots || {};
   const { markupBefore = '', markupAfter = '' } = other || {};
@@ -59,7 +61,11 @@ const initBasicFlyout = (
   ${[header, content, footer, subFooter].filter(Boolean).join('\n  ')}
 </p-flyout>`;
 
-  return setContentWithDesignSystem(page, [markupBefore, flyoutMarkup, markupAfter].filter(Boolean).join('\n'));
+  return setContentWithDesignSystem(
+    page,
+    [markupBefore, flyoutMarkup, markupAfter].filter(Boolean).join('\n'),
+    options
+  );
 };
 
 const initAdvancedFlyout = async (page: Page) => {
@@ -793,5 +799,53 @@ test.describe('after dynamic slot change', () => {
     await page.setViewportSize({ width: 320, height: 500 });
 
     expect(await getStickyTopCssVarValue(page)).toBe('167px');
+  });
+});
+
+test.describe('events', () => {
+  skipInBrowsers(['firefox', 'webkit']);
+  test('should call motionVisibleEnd event when opening transition is finished', async ({ page }) => {
+    await initBasicFlyout(
+      page,
+      { open: false },
+      {},
+      {},
+      { injectIntoHead: '<style>:root { --p-transition-duration: unset; }</style>' }
+    );
+    const host = await getHost(page);
+    await waitForStencilLifecycle(page);
+    await addEventListener(host, 'motionVisibleEnd');
+    await addEventListener(host, 'motionHiddenEnd');
+
+    expect((await getEventSummary(host, 'motionVisibleEnd')).counter).toBe(0);
+    expect((await getEventSummary(host, 'motionHiddenEnd')).counter).toBe(0);
+
+    await openFlyout(page);
+    await waitForFlyoutTransition();
+
+    expect((await getEventSummary(host, 'motionVisibleEnd')).counter).toBe(1);
+    expect((await getEventSummary(host, 'motionHiddenEnd')).counter).toBe(0);
+  });
+  test('should call motionHiddenEnd event when closing transition is finished', async ({ page }) => {
+    await initBasicFlyout(
+      page,
+      { open: true },
+      {},
+      {},
+      { injectIntoHead: '<style>:root { --p-transition-duration: unset; }</style>' }
+    );
+    const host = await getHost(page);
+    await waitForStencilLifecycle(page);
+    await addEventListener(host, 'motionVisibleEnd');
+    await addEventListener(host, 'motionHiddenEnd');
+
+    expect((await getEventSummary(host, 'motionVisibleEnd')).counter).toBe(0);
+    expect((await getEventSummary(host, 'motionHiddenEnd')).counter).toBe(0);
+
+    await dismissFlyout(page);
+    await waitForFlyoutTransition();
+
+    expect((await getEventSummary(host, 'motionVisibleEnd')).counter).toBe(0);
+    expect((await getEventSummary(host, 'motionHiddenEnd')).counter).toBe(1);
   });
 });