+++ packages/components-js/tests/e2e/specs/modal.e2e.ts
@@ -9,16 +9,16 @@ import {
   getElementStyle,
   getEventSummary,
   getLifecycleStatus,
-  getProperty,
   setContentWithDesignSystem,
   setProperty,
   skipInBrowsers,
   sleep,
   waitForStencilLifecycle,
+  type Options,
 } from '../helpers';
 import type { ModalAriaAttribute, SelectedAriaAttributes } from '@porsche-design-system/components';
 
-const CSS_TRANSITION_DURATION = 600;
+const CSS_TRANSITION_DURATION = 600; // Corresponds to motionDurationLong
 
 const getHost = (page: Page) => page.locator('p-modal');
 const getScrollContainer = (page: Page) => page.locator('p-modal .scroller');
@@ -29,6 +29,7 @@ const getDismissButton = (page: Page) => page.locator('p-modal p-button-pure.dis
 const getFooter = (page: Page) => page.locator('p-modal slot[name="footer"]');
 const getFooterBoxShadow = async (page: Page): Promise<string> => getElementStyle(getFooter(page), 'boxShadow');
 const getBodyStyle = async (page: Page) => getAttribute(page.locator('body'), 'style');
+const waitForModalTransition = async () => sleep(CSS_TRANSITION_DURATION);
 
 const initBasicModal = (
   page: Page,
@@ -43,7 +44,8 @@ const initBasicModal = (
     disableCloseButton?: boolean;
     markupBefore?: string;
     markupAfter?: string;
-  }
+  },
+  options?: Options
 ): Promise<void> => {
   const {
     isOpen = true,
@@ -74,7 +76,8 @@ const initBasicModal = (
   ${hasSlottedHeader ? '<div slot="header"><h2>Some Heading</h2><p>Some header content</p></div>' : ''}
   ${content}
   ${hasSlottedFooter ? '<div slot="footer">Some Footer</div>' : ''}
-</p-modal>${markupAfter ? markupAfter : ''}`
+</p-modal>${markupAfter ? markupAfter : ''}`,
+    options
   );
 };
 
@@ -820,5 +823,49 @@ test.describe('after dynamic slot change', () => {
 
     await expect(page.getByText(footerText)).toBeVisible();
     expect(await getFooterBoxShadow(page)).toBe('rgba(204, 204, 204, 0.35) 0px -5px 10px 0px');
+  });
+});
+
+test.describe('events', () => {
+  skipInBrowsers(['firefox', 'webkit']);
+  test('should call motionVisibleEnd event when opening transition is finished', async ({ page }) => {
+    await initBasicModal(
+      page,
+      { isOpen: false },
+      { injectIntoHead: '<style>:root { --p-transition-duration: unset; }</style>' }
+    );
+    const host = getHost(page);
+    await waitForStencilLifecycle(page);
+    await addEventListener(host, 'motionVisibleEnd');
+    await addEventListener(host, 'motionHiddenEnd');
+
+    expect((await getEventSummary(host, 'motionVisibleEnd')).counter).toBe(0);
+    expect((await getEventSummary(host, 'motionHiddenEnd')).counter).toBe(0);
+
+    await openModal(page);
+    await waitForModalTransition();
+
+    expect((await getEventSummary(host, 'motionVisibleEnd')).counter).toBe(1);
+    expect((await getEventSummary(host, 'motionHiddenEnd')).counter).toBe(0);
+  });
+  test('should call motionHiddenEnd event when closing transition is finished', async ({ page }) => {
+    await initBasicModal(
+      page,
+      { isOpen: true },
+      { injectIntoHead: '<style>:root { --p-transition-duration: unset; }</style>' }
+    );
+    const host = getHost(page);
+    await waitForStencilLifecycle(page);
+    await addEventListener(host, 'motionVisibleEnd');
+    await addEventListener(host, 'motionHiddenEnd');
+
+    expect((await getEventSummary(host, 'motionVisibleEnd')).counter).toBe(0);
+    expect((await getEventSummary(host, 'motionHiddenEnd')).counter).toBe(0);
+
+    await dismissModal(page);
+    await waitForModalTransition();
+
+    expect((await getEventSummary(host, 'motionVisibleEnd')).counter).toBe(0);
+    expect((await getEventSummary(host, 'motionHiddenEnd')).counter).toBe(1);
   });
 });