+++ packages/components/src/components/carousel/carousel.tsx
@@ -16,6 +16,7 @@ import {
   getAmountOfPages,
   getSlidesAndAddAttributes,
   getSplideBreakpoints,
+  getLangDirection,
   isInfinitePagination,
   renderPagination,
   slideNext,
@@ -217,6 +218,7 @@ export class Carousel {
       breakpoints: getSplideBreakpoints(this.slidesPerPage as Exclude<BreakpointCustomizable<number> | 'auto', string>), // eslint-disable-line @typescript-eslint/no-redundant-type-constituents
       // https://splidejs.com/guides/i18n/#default-texts
       i18n: parseJSONAttribute(this.intl || {}), // can only be applied initially atm
+      direction: getLangDirection(this.host),
     });
 
     this.registerSplideHandlers(this.splide);
@@ -279,7 +281,8 @@ export class Carousel {
         CarouselAlignHeader,
         CarouselAlignHeaderDeprecated
       >,
-      this.theme
+      this.theme,
+      this.hasNavigation
     );
 
     const PrefixedTagNames = getPrefixedTagNames(this.host);